Move some methods to private extension
