+++ catowseriOS/CoreBrowser/Tabs/TabsListError.swift
@@ -1,0 +1,16 @@
+//
+//  TabsListError.swift
+//  CoreBrowser
+//
+//  Created by Andrey Ermoshin on 25.07.2023.
+//  Copyright © 2023 andreiermoshin. All rights reserved.
+//
+
+import Foundation
+
+public enum TabsListError: LocalizedError {
+    case notInitializedYet
+    case selectedNotFound
+    case wrongTabContent
+    case wrongTabIndexToReplace
+}
